+++ release/i18next-1.6.1pre.js
@@ -891,7 +891,11 @@
             
             translated = translate(pluralKey, optionWithoutCount);
             if (translated != o.pluralNotFound) {
-                return applyReplacement(translated, { count: options.count }); // apply replacement for count only
+                return applyReplacement(translated, {
+                    count: options.count,
+                    interpolationPrefix: options.interpolationPrefix,
+                    interpolationSuffix: options.interpolationSuffix 
+                }); // apply replacement for count only
             } // else continue translation with original/singular key
         }